bash_completion: small optimization
Right now we always call "hg help $cmd" to get the canonical name of $cmd
(i.e. to go from "co" to "update").

This patch optimistically assumes that $cmd is already the canonical form
and tries to generate completions for it.  If that fails, it falls back
to canonicalizing $cmd and trying again.

This means that:

- if a command or alias is explicitly handled by the
  _hg_command_specific function, things get somewhat faster

- as long as the canonical $cmd is handled by _hg_command_specific, all
  its aliases and abbreviations are also handled.