Always copy the necessary files before applying a git patch
This patch removes the "copymod" attribute from the gitpatch
class.

AFAICS, that attribute was only used to delay the copying of
renamed/copied files if there are no other changes to the target,
but in this case, if there are changes to the source, we'll end
up copying the wrong version.

This should fix issue762.