rebase: add --detach option to detach intermediate revisions (issue1950)
When rebasing an intermediate revision, rebase keeps a parent relationship
with the original parent. This option forces the removal of this relationship.
In more depth, it 'fakes' null merges between the target revision and the
ancestors of source, dropping every change from the ancestors.
The result is that every change in source and its descendants will be rebased,
ignoring the changes in its ancestors.

--HG--
extra : rebase_source : d3d746a7ded685fefae1378c6bbb7c7d084170f1