heads: mention STARTREV in synopsis and help for --rev
Before this change, "hg help heads" said

  hg heads [-ac] [-r REV] [REV]...

      [...]
      If STARTREV is specified, only those heads that are descendants
      of STARTREV will be displayed.
      [...]

   -r --rev REV       show only heads which are descendants of REV
   [...]

which made little sense since there are two things called REV in the
synopsis and nothing called STARTREV.

A little digging reveals that the "[-r REV]" part of the synopsis was
introduced in 8e503fa54d2d, changed to "[-r STARTREV]" in
2bcef677a6c3, and then changed back to "[-r REV]" in 4b02fc71bbba.

The last change seems to be based on a patch[1] on our mailinglist
that actually *inserted* STARTREV again in the help for the command
line option itself. For some reason, the patch was changed to remove
STARTREV from the synopsis.

This change finally makes the help consistent by putting STARTREV back
into the help in all places where it is needed:

  hg heads [-ac] [-r STARTREV] [REV]...

      [...]
      If STARTREV is specified, only those heads that are descendants
      of STARTREV will be displayed.
      [...]

   -r --rev STARTREV  show only heads which are descendants of STARTREV
   [...]

This was not possible until 40c06bbf58be, which introduced the
possibility of naming the meta variables for each option.

[1]: http://mercurial.markmail.org/message/qgc55gd4fam4ogvz