peer: subrepo isolation, pass repo instead of repo.ui to hg.peer
Do not pass ui because it contains the configuration of the repo. It is the
same object as repo.ui.
When a repo is passed to hg.peer, the global configuration is read from
repo.baseui.

--HG--
branch : stable
extra : rebase_source : ee40dab7fb1749f0a3739cb04e0202019ed5f379