contrib/Makefile.python: build local Python from source
This provides an easy way to install new or old Python versions on Unixish
systems. It is mainly intended for testing Mercurial with different Python
versions.

Example:
  $ cd build
  $ make -f ../contrib/Makefile.python python PYTHONVER=2.4 PREFIX=/tmp/p24
  $ /tmp/p24/bin/python -V
  Python 2.4

Tested on Fedora, Debian and Ubuntu with Python versions from 2.4 to 2.7.6.

Most Python versions need some workarounds to compile on Debian-ish systems.
The workarounds do not do any significant harm on Fedora.