histedit-test: replace obscure python script by a plain file
One of the tests uses a python script to edit the histedit commands on
the fly. It was hard to read and likely to break. This is replaced by
a plain command file.

Doing so made me discover that the python script is incorrect. The
"+6" regexp never matched because the revision number is wrong.

This error is kept in the replacement for now. Fixing the histedit command
triggers a traceback in histedit.

A later changeset will fix the error and restore the intended test.

--HG--
extra : rebase_source : 7076af979fe772083aaa71e07a3f5f7361660afc