test-convert-hg-startrev: fix test for copy removal on missing parent
The test assumed that 'b' was missing just because the changeset that
introduced 'b' wasn't converted ... but 'b' was also not removed before the
start revision, and the parent was thus not missing at all.

Instead we introduce a file 'f' in rev 0 and remove it in rev 1 so the copy
source really doesn't exist in the repo that is converted from rev 1.

--HG--
extra : rebase_source : ec577759c30e2f2f9aa38b43fbf22cb828f39803