record: --user/-u now works with record when ui.username not set (issue3857)
The -u flag didn't work when ui.username was not set and resulted in an
abort message. This was fixed by checking for the 'user' key in the opts
dictionary. If the key is present, the step causing the exception is not
executed.

--HG--
extra : rebase_source : c9132c76370b159a35a8b4422aa0b909ba6457b6