branchmap: simplify update code
We drop iterrevs which are not needed anymore. The know head are never a
descendant of the updated set. It was possible with the old strip code. This
simplification make the code easier to read an update.

--HG--
extra : rebase_source : a365a42fe575cd3a0b726861fc56c1d6ab89e511