bundle2: support chunk iterator as part data
When the `part.data` attribute is an iterator, we assume it is an iterator of
chunks and use it.

We use a chunkbuffer to yield chunks of 4096 bytes.

The tests are updated to use this feature.

--HG--
extra : rebase_source : b14cb7fa7e035af295eeb21416f7d63c988baf0d
extra : amend_source : 04e14bc1567f04ffc26fbaf5d66760cc228c0a7d