bundle2: capture remote stdout while unbundling
When a reply is built, the bundle processing will capture the output of each
handler and sends it to the client in a dedicated part.

As a side effect, this add a "remote: " prefix to destination output on local
push. This is considered okay for now as:

1. bundle2 is still experimental,
2. Matt said he could be okay to change output for bundle2,
3. This keeps the implementation simple.

This changeset does it for stdout only. stderr will be done in a future changeset.

--HG--
extra : rebase_source : 91e23107fc5747144c96573aa75fe763b6e8f7a4
extra : amend_source : 428dff5786a95133aadbeb2d07a4d74f1d2cddfe